- Networking: PlayerLines objects

+++ LineFighter/Assets/Scripts/Interfaces/ILine.cs
@@ -2,7 +2,7 @@
 {
     bool LineGravity { get; set; }
 
-    float Thickness { get; set; }
+    double Thickness { get; set; }
 
     // Per 1/10 second
     int RefillRate { get; set; }
